requirements.txt
python-dotenv==1.0.1
scratchattach==1.6.2
requests==2.31.0
beautifulsoup4==4.12.3
